docker/lockmancyber/resnet_classify/requirements.txt
numpy<2.0
torch==2.2.2
torchvision==0.17.2
flask
flask-cors
pillow